+++ app/controllers/home_controller.rb
@@ -26,14 +26,14 @@ class HomeController < ApplicationController
   end
 
   def toggle_mobile
-    if session[:mobile_view].nil?
-      # we're most probably not on mobile, but user wants it anyway
-      session[:mobile_view] = true
-    else
-      # switch from mobile to normal html
-      session[:mobile_view] = !session[:mobile_view]
-    end
+    session[:mobile_view] = session[:mobile_view].nil? ? true : !session[:mobile_view]
 
     redirect_to :back
+  end
+
+  def force_mobile
+    session[:mobile_view] = true
+
+    redirect_to stream_path
   end
 end