+++ app/models/poll_participation.rb
@@ -1,7 +1,7 @@
 class PollParticipation < ActiveRecord::Base
 
   include Diaspora::Federated::Base
-  
+
   include Diaspora::Guid
   include Diaspora::Relayable
   belongs_to :poll
@@ -37,7 +37,7 @@ class PollParticipation < ActiveRecord::Base
   end
 
   def diaspora_handle= nh
-    self.author = Webfinger.new(nh).fetch
+    self.author = Person.find_or_fetch_by_identifier(nh)
   end
 
   def not_already_participated